\name{Mobius-package}
\alias{Mobius-package}
\docType{package}
\title{
Mobius Transport for Directional Data
}

\description{
Description: A multivariate zero inflated logistic normal regression model
is implemented for compositional data with zero values present.
}

\details{
\tabular{ll}{
Package: \tab Mobius\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2026-08-21 \cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>.
}

\references{
Li Z., Lee K., Karagas M. R., Madan J. C., Hoen A. G., O'malley A. J. and Li H. (2018).
Conditional regression based on a multivariate zero-inflated logistic-normal model for
microbiome relative abundance data.
Statistics in Biosciences, 10(3): 587--608.
}





